% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_tem.R
\name{predator_prey_T}
\alias{predator_prey_T}
\title{predator-prey ODE with T dependence but without size dependence(type I)
N is the biomass of prey and P is the biomass of predator}
\usage{
predator_prey_T(t, state, parms)
}
\description{
predator-prey ODE with T dependence but without size dependence(type I)
N is the biomass of prey and P is the biomass of predator
}
