% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_orig.R
\name{predator_prey_type1}
\alias{predator_prey_type1}
\title{original predator-prey ODE without T and size dependence(type I)
N is the biomass of prey and P is the biomass of predator}
\usage{
predator_prey_type1(t, state, parms)
}
\description{
original predator-prey ODE without T and size dependence(type I)
N is the biomass of prey and P is the biomass of predator
}
